Rename files and components
