+++ renderer/components/icons/sidebar-closed-svg.tsx
