+++ renderer/components/icons/sidebar-opened-svg.tsx
