+++ renderer/components/icons/upscayl-logo-svg.tsx
