+++ renderer/components/main-content/progress-bar.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from "react";
-import UpscaylSVGLogo from "@/components/icons/Logo";
+import UpscaylSVGLogo from "@/components/icons/upscayl-logo-svg";
 import { useAtomValue } from "jotai";
 import { translationAtom } from "@/atoms/translations-atom";
 import ELECTRON_COMMANDS from "@common/commands";
