+++ renderer/components/main-content/reset-button.tsx
