+++ renderer/components/sidebar/settings-tab/donate-button.tsx
