+++ renderer/components/sidebar/settings-tab/index.tsx
@@ -9,12 +9,12 @@ import { DonateButton } from "./donate-button";
 import React, { useEffect, useState } from "react";
 import { themeChange } from "theme-change";
 import { useAtom, useAtomValue } from "jotai";
-import { customModelsPathAtom, scaleAtom } from "../../atoms/userSettingsAtom";
-import { modelsListAtom } from "../../atoms/modelsListAtom";
-import useLogger from "../hooks/use-logger";
+import { customModelsPathAtom, scaleAtom } from "@/atoms/user-settings-atom";
+import { modelsListAtom } from "@/atoms/models-list-atom";
+import useLogger from "@/components/hooks/use-logger";
 import { InputCompression } from "./input-compression";
 import OverwriteToggle from "./overwrite-toggle";
-import { UpscaylCloudModal } from "../UpscaylCloudModal";
+import { UpscaylCloudModal } from "@/components/upscayl-cloud-modal";
 import { ResetSettingsButton } from "./reset-settings-button";
 import { FEATURE_FLAGS } from "@common/feature-flags";
 import TurnOffNotificationsToggle from "./turn-off-notifications-toggle";
