+++ renderer/components/sidebar/settings-tab/input-compression.tsx
