+++ renderer/components/sidebar/settings-tab/input-gpu-id.tsx
