+++ renderer/components/sidebar/settings-tab/log-area.tsx
