+++ renderer/components/sidebar/settings-tab/reset-settings-button.tsx
