+++ renderer/components/sidebar/settings-tab/select-image-format.tsx
