+++ renderer/components/sidebar/settings-tab/select-theme.tsx
